Backend-Proyecto2/requirements.txt
fastapi
uvicorn
python-dotenv
redis
passlib[bcrypt]
python-jose[cryptography]